#1180 Calculate total number of entries and total size of a folder's tree

+++ logicaldoc-gui/src/main/java/com/logicaldoc/gui/frontend/client/document/grid/DocumentsListGrid.java
@@ -1113,7 +1113,7 @@ public class DocumentsListGrid extends RefreshableListGrid implements DocumentsG
 		if (getGridCursor() != null) {
 			getGridCursor().setPageSize(pageSize);
 			if (folder != null)
-				getGridCursor().setTotalRecords(folder.getDocumentCount());
+				getGridCursor().setTotalRecords((int)folder.getDocumentCount());
 		}
 
 		/*
